auth-service/requirements.txt
Django==4.2.7
djangorestframework==3.14.0
djangorestframework-simplejwt==5.3.0
PyJWT==2.8.0
PyMySQL==1.1.2
django-cors-headers==4.3.0
python-decouple==3.8